PO-1721, PO-1683. Add ct.gov to batch upload. Refactor Excel Loader. Peer Review Christophe
git-svn-id: https://ncisvn.nci.nih.gov/svn/coppa/trunk/code@9134 553898ca-32d5-487a-af8b-6cf4ca579ab4

+++ pa/pa-ejb/src/java/gov/nih/nci/pa/service/TrialRegistrationBeanLocal.java
@@ -701,35 +701,36 @@ public class TrialRegistrationBeanLocal implements TrialRegistrationServiceLocal
         ssSourceDTO.setStudyProtocolIdentifier(targetSpIi);
         studySiteService.delete(sourceIi);
         studySiteService.update(ssSourceDTO);
-        //sponsor
-        studySiteDto = new StudySiteDTO();
-        studySiteDto.setStudyProtocolIdentifier(sourceSpIi);
-        studySiteDto.setFunctionalCode(CdConverter.convertToCd(StudySiteFunctionalCode.SPONSOR));
+        if (target.getCtgovXmlRequiredIndicator().booleanValue()) {
+           //sponsor
+           studySiteDto = new StudySiteDTO();
+           studySiteDto.setStudyProtocolIdentifier(sourceSpIi);
+           studySiteDto.setFunctionalCode(CdConverter.convertToCd(StudySiteFunctionalCode.SPONSOR));
 
-        List<StudySiteDTO> studySiteSponsorDtos =
-            paServiceUtils.getStudySite(studySiteDto, true);
-        StudySiteDTO ssSponsorSourceDTO = null;
-        if (PAUtil.getFirstObj(studySiteSponsorDtos) != null) {
-            ssSponsorSourceDTO = PAUtil.getFirstObj(studySiteSponsorDtos);
-        } else {
-            throw new PAException("Source Lead Organization is not available");
+           List<StudySiteDTO> studySiteSponsorDtos =
+              paServiceUtils.getStudySite(studySiteDto, true);
+           StudySiteDTO ssSponsorSourceDTO = null;
+           if (PAUtil.getFirstObj(studySiteSponsorDtos) != null) {
+              ssSponsorSourceDTO = PAUtil.getFirstObj(studySiteSponsorDtos);
+           } else {
+              throw new PAException("Source Lead Organization is not available");
+          }
+          sourceIi = ssSponsorSourceDTO.getIdentifier();
+          studySiteDto = new StudySiteDTO();
+          studySiteDto.setStudyProtocolIdentifier(targetSpIi);
+          studySiteDto.setFunctionalCode(CdConverter.convertToCd(StudySiteFunctionalCode.SPONSOR));
+          studySiteSponsorDtos = paServiceUtils.getStudySite(studySiteDto, true);
+          StudySiteDTO ssSponsorTargetDTO = null;
+          if (PAUtil.getFirstObj(studySiteSponsorDtos) != null) {
+             ssSponsorTargetDTO = PAUtil.getFirstObj(studySiteSponsorDtos);
+          } else {
+             throw new PAException("Target Sponsor is not available");
+          }
+          ssSponsorSourceDTO.setIdentifier(ssSponsorTargetDTO.getIdentifier());
+          ssSponsorSourceDTO.setStudyProtocolIdentifier(targetSpIi);
+          studySiteService.delete(sourceIi);
+          studySiteService.update(ssSponsorSourceDTO);
         }
-        sourceIi = ssSponsorSourceDTO.getIdentifier();
-        studySiteDto = new StudySiteDTO();
-        studySiteDto.setStudyProtocolIdentifier(targetSpIi);
-        studySiteDto.setFunctionalCode(CdConverter.convertToCd(StudySiteFunctionalCode.SPONSOR));
-        studySiteSponsorDtos = paServiceUtils.getStudySite(studySiteDto, true);
-        StudySiteDTO ssSponsorTargetDTO = null;
-        if (PAUtil.getFirstObj(studySiteSponsorDtos) != null) {
-            ssSponsorTargetDTO = PAUtil.getFirstObj(studySiteSponsorDtos);
-        } else {
-            throw new PAException("Target Sponsor is not available");
-        }
-        ssSponsorSourceDTO.setIdentifier(ssSponsorTargetDTO.getIdentifier());
-        ssSponsorSourceDTO.setStudyProtocolIdentifier(targetSpIi);
-        studySiteService.delete(sourceIi);
-        studySiteService.update(ssSponsorSourceDTO);
-
         paServiceUtils.executeSql(deleteAndReplace(sourceSpIi , targetSpIi));
       }
     } catch (Exception e) {
